requirements.txt
Flask
requests
pandas
pandas-ta
numpy
scikit-learn
tensorflow
google-generativeai
